requirements.txt
flask==3.1.0
pandas==2.2.3
google-cloud-storage==3.1.0
google-cloud-firestore==2.20.2
pytest